+++ modApiServer/test/org/aion/api/server/ApiTest.java
@@ -107,6 +107,10 @@ public class ApiTest {
     public void tearDown() {
         // get a list of all the files in keystore directory
         File folder = new File(KEYSTORE_PATH);
+
+        if (folder == null)
+            return;
+
         File[] AllFilesInDirectory = folder.listFiles();
 
         // check for invalid or wrong path - should not happen
@@ -118,7 +122,13 @@ public class ApiTest {
                 file.delete();
         }
         folder = new File(DATABASE_PATH);
+        if (folder == null)
+            return;
         AllFilesInDirectory = folder.listFiles();
+
+        if (AllFilesInDirectory == null)
+            return;
+
         for (File file : AllFilesInDirectory) {
             file.delete();
         }
