+++ modApiServer/test/org/aion/api/server/pb/ApiAion0Test.java
@@ -131,6 +131,10 @@ public class ApiAion0Test {
 
         // get a list of all the files in keystore directory
         File folder = new File(KEYSTORE_PATH);
+
+        if (folder == null)
+            return;
+
         File[] AllFilesInDirectory = folder.listFiles();
 
         // check for invalid or wrong path - should not happen
@@ -142,7 +146,13 @@ public class ApiAion0Test {
                 file.delete();
         }
         folder = new File(DATABASE_PATH);
+        if (folder == null)
+            return;
         AllFilesInDirectory = folder.listFiles();
+
+        if (AllFilesInDirectory == null)
+            return;
+
         for (File file : AllFilesInDirectory) {
             file.delete();
         }
